+++ package.json
@@ -3,6 +3,8 @@
   "version": "1.0.0",
   "description": "> This project is a simple HTML list of To Do tasks.",
   "devDependencies": {
+    "@babel/plugin-transform-modules-commonjs": "^7.18.6",
+    "@babel/preset-env": "^7.18.10",
     "babel-eslint": "^10.1.0",
     "css-loader": "^6.7.1",
     "csv-loader": "^3.0.5",
@@ -11,6 +13,7 @@
     "eslint-plugin-import": "^2.26.0",
     "hint": "^7.1.1",
     "html-webpack-plugin": "^5.5.0",
+    "jest": "^28.1.3",
     "style-loader": "^3.3.1",
     "webpack": "^5.74.0",
     "webpack-cli": "^4.10.0",
@@ -18,11 +21,14 @@
     "xml-loader": "^1.2.1"
   },
   "dependencies": {
+    "babel-jest": "^28.1.3",
     "gh-pages": "^4.0.0",
-    "lodash": "^4.17.21"
+    "jest-environment-jsdom": "^28.1.3",
+    "lodash": "^4.17.21",
+    "ts-jest": "^28.0.8"
   },
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "jest",
     "watch": "webpack --watch",
     "start": "webpack serve --open",
     "build": "webpack",
