+++ src/index.js
@@ -1,2 +1,3 @@
 import './todo.js';
 import './style.css';
+import './todoEvent.js'
