+++ src/todo.js
@@ -1,9 +1,6 @@
 import updateStatus from './status.js';
 
 const listContainer = document.querySelector('.lists');
-const todoInput = document.querySelector('.todo-input');
-const addTodoBtn = document.querySelector('#todo-add-btn');
-const clearAllCompletedBtn = document.querySelector('.clear-all-completed p');
 
 export default class Todos {
   constructor(tasks = []) {
@@ -148,32 +145,3 @@ export default class Todos {
     this.getTasks();
   }
 }
-
-const todos = new Todos();
-
-todos.getTasks();
-
-addTodoBtn.addEventListener('click', (e) => {
-  e.preventDefault();
-  const index = todos.tasks.length + 1;
-  if (todoInput.value !== '') {
-    const newTask = {
-      description: todoInput.value,
-      completed: false,
-      index,
-    };
-    todos.addTask(newTask);
-    todoInput.value = '';
-  }
-});
-
-todoInput.addEventListener('keypress', (e) => {
-  if (e.key === 'Enter') {
-    e.preventDefault();
-    addTodoBtn.click();
-  }
-});
-
-clearAllCompletedBtn.addEventListener('click', () => {
-  todos.clearAllCompleted();
-});
